Add Python 3 support
This commit adds Python 3.3 and 3.4 support. It was originally gonna
drop Python 2.6 support, but it can easily be supported.

The biggest influence in the diff is changes to absolute imports and
dropping the explicit use of unittest2, since imports changed with
python 3 and unittest2 is no longer needed.

The small changes to `setup.py` and `setup.cfg` are related and all were
needed for the code to all work, so I thought it fit better within a
single commit.

I added `six` as a requirement since it made using `iteritems` still
work on dictionary items.

I changed from `pinocchio` to `spec` since it was the only used part of
`pinocchio` and it was the only one that had Python 3 support when I
first looked.

The requirement for `six` means that `querylist` can no longer be
imported easily in `setup.py` without the outer virtualenv/python having
six installed, and the only value that changed often was the version,
so I used the code @mitsuhiko's `click` uses for reading the
`__version__` from the `__init__.py` file.

I also dropped the requirement for `unittest2py3k` since it isn't
needed by Python 3 anymore, and I changed the `assertEquals` to
`assertEqual` since the former is being deprecated, which is the only
change that could've probably gone in a separate commit.